+++ test/integration/components.cc
@@ -17,6 +17,7 @@
 
 #include <gtest/gtest.h>
 
+#include <sdf/Cylinder.hh>
 #include <sdf/Element.hh>
 
 #include "ignition/gazebo/components/Altimeter.hh"
@@ -46,6 +47,7 @@
 #include "ignition/gazebo/components/ParentEntity.hh"
 #include "ignition/gazebo/components/ParentLinkName.hh"
 #include "ignition/gazebo/components/Performer.hh"
+#include "ignition/gazebo/components/PerformerLevels.hh"
 #include "ignition/gazebo/components/Pose.hh"
 #include "ignition/gazebo/components/Sensor.hh"
 #include "ignition/gazebo/components/Static.hh"
@@ -212,6 +214,12 @@ TEST_F(ComponentsTest, Collision)
 TEST_F(ComponentsTest, Geometry)
 {
   auto data1 = sdf::Geometry();
+  data1.SetType(sdf::GeometryType::CYLINDER);
+  sdf::Cylinder cylinderShape;
+  cylinderShape.SetRadius(1.23);
+  cylinderShape.SetLength(4.56);
+  data1.SetCylinderShape(cylinderShape);
+
   auto data2 = sdf::Geometry();
 
   // Create components
@@ -219,7 +227,18 @@ TEST_F(ComponentsTest, Geometry)
   auto comp12 = components::Geometry(data1);
   auto comp2 = components::Geometry(data2);
 
-  // TODO(anyone) Stream operator
+  // TODO(anyone) Equality operators
+
+  // Stream operators
+  std::ostringstream ostr;
+  ostr << comp11;
+  std::istringstream istr(ostr.str());
+  components::Geometry comp3;
+  istr >> comp3;
+  EXPECT_EQ(sdf::GeometryType::CYLINDER, comp3.Data().Type());
+  ASSERT_NE(nullptr, comp3.Data().CylinderShape());
+  EXPECT_DOUBLE_EQ(1.23, comp3.Data().CylinderShape()->Radius());
+  EXPECT_DOUBLE_EQ(4.56, comp3.Data().CylinderShape()->Length());
 }
 
 /////////////////////////////////////////////////
@@ -290,7 +309,13 @@ TEST_F(ComponentsTest, Inertial)
   EXPECT_FALSE(comp11 == comp2);
   EXPECT_FALSE(comp11 != comp12);
 
-  // TODO(anyone) Stream operator
+  // Stream operators
+  std::ostringstream ostr;
+  ostr << comp11;
+  std::istringstream istr(ostr.str());
+  components::Inertial comp3;
+  istr >> comp3;
+  EXPECT_DOUBLE_EQ(1.0, comp3.Data().MassMatrix().Mass());
 }
 
 /////////////////////////////////////////////////
@@ -311,7 +336,7 @@ TEST_F(ComponentsTest, Joint)
   EXPECT_TRUE(ostr.str().empty());
 
   std::istringstream istr("ignored");
-  components::CanonicalLink comp3;
+  components::Joint comp3;
   istr >> comp3;
 }
 
@@ -319,6 +344,15 @@ TEST_F(ComponentsTest, Joint)
 TEST_F(ComponentsTest, JointAxis)
 {
   auto data1 = sdf::JointAxis();
+  data1.SetXyz(math::Vector3d(1, 2, 3));
+  data1.SetUseParentModelFrame(true);
+  data1.SetDamping(0.1);
+  data1.SetFriction(0.2);
+  data1.SetLower(0.3);
+  data1.SetUpper(0.4);
+  data1.SetEffort(0.5);
+  data1.SetMaxVelocity(0.6);
+
   auto data2 = sdf::JointAxis();
 
   // Create components
@@ -326,21 +360,56 @@ TEST_F(ComponentsTest, JointAxis)
   auto comp12 = components::JointAxis(data1);
   auto comp2 = components::JointAxis(data2);
 
-  // TODO(anyone) Stream operator
+  // TODO(anyone) Equality operators
+
+  // Stream operators
+  std::ostringstream ostr;
+  ostr << comp11;
+  std::istringstream istr(ostr.str());
+
+  components::JointAxis comp3;
+  istr >> comp3;
+
+  EXPECT_EQ(math::Vector3d(1, 2, 3), comp3.Data().Xyz());
+  EXPECT_DOUBLE_EQ(0.1, comp3.Data().Damping());
+  EXPECT_DOUBLE_EQ(0.2, comp3.Data().Friction());
+  EXPECT_DOUBLE_EQ(0.3, comp3.Data().Lower());
+  EXPECT_DOUBLE_EQ(0.4, comp3.Data().Upper());
+  EXPECT_DOUBLE_EQ(0.5, comp3.Data().Effort());
+  EXPECT_DOUBLE_EQ(0.6, comp3.Data().MaxVelocity());
+  EXPECT_TRUE(comp3.Data().UseParentModelFrame());
 }
 
 /////////////////////////////////////////////////
 TEST_F(ComponentsTest, JointType)
 {
-  auto data1 = sdf::JointType();
-  auto data2 = sdf::JointType();
+  auto data1 = sdf::JointType::FIXED;
+  auto data2 = sdf::JointType::REVOLUTE;
 
   // Create components
   auto comp11 = components::JointType(data1);
   auto comp12 = components::JointType(data1);
   auto comp2 = components::JointType(data2);
 
-  // TODO(anyone) Stream operator
+  // Equality operators
+  EXPECT_EQ(comp11, comp12);
+  EXPECT_NE(comp11, comp2);
+  EXPECT_TRUE(comp11 == comp12);
+  EXPECT_TRUE(comp11 != comp2);
+  EXPECT_FALSE(comp11 == comp2);
+  EXPECT_FALSE(comp11 != comp12);
+
+  // Stream operators
+  std::ostringstream ostr;
+  ostr << comp11;
+  EXPECT_EQ(std::to_string(static_cast<int>(sdf::JointType::FIXED)),
+      ostr.str());
+
+  std::istringstream istr(std::to_string(static_cast<int>(
+      sdf::JointType::SCREW)));
+  components::JointType comp3;
+  istr >> comp3;
+  EXPECT_EQ(sdf::JointType::SCREW, comp3.Data());
 }
 
 /////////////////////////////////////////////////
@@ -529,6 +598,7 @@ TEST_F(ComponentsTest, Link)
 TEST_F(ComponentsTest, Material)
 {
   auto data1 = sdf::Material();
+  data1.SetAmbient(math::Color(1, 0, 0, 1));
   auto data2 = sdf::Material();
 
   // Create components
@@ -536,7 +606,13 @@ TEST_F(ComponentsTest, Material)
   auto comp12 = components::Material(data1);
   auto comp2 = components::Material(data2);
 
-  // TODO(anyone) Stream operator
+  // Stream operators
+  std::ostringstream ostr;
+  ostr << comp11;
+  std::istringstream istr(ostr.str());
+  components::Material comp3;
+  istr >> comp3;
+  EXPECT_EQ(math::Color(1, 0, 0, 1), comp3.Data().Ambient());
 }
 
 /////////////////////////////////////////////////
@@ -662,6 +738,32 @@ TEST_F(ComponentsTest, Performer)
   std::istringstream istr("ignored");
   components::Performer comp3;
   istr >> comp3;
+}
+
+/////////////////////////////////////////////////
+TEST_F(ComponentsTest, PerformerLevels)
+{
+  // Create components
+  auto comp11 = components::PerformerLevels({1, 2, 3});
+  auto comp12 = components::PerformerLevels({1, 2, 3});
+  auto comp2 = components::PerformerLevels({4, 5});
+
+  // TODO(anyone) Equality operators
+
+  // Stream operators
+  std::ostringstream ostr;
+  ostr << comp11;
+  EXPECT_EQ("1 2 3 ", ostr.str());
+
+  std::istringstream istr("7 8 9 10");
+  components::PerformerLevels comp3;
+  istr >> comp3;
+  EXPECT_EQ(4u, comp3.Data().size());
+  EXPECT_EQ(1u, comp3.Data().count(7));
+  EXPECT_EQ(1u, comp3.Data().count(8));
+  EXPECT_EQ(1u, comp3.Data().count(9));
+  EXPECT_EQ(1u, comp3.Data().count(10));
+  EXPECT_EQ(0u, comp3.Data().count(1));
 }
 
 /////////////////////////////////////////////////